IOS-4979 Fix review
