+++ Tangem/Common/UI/ProviderRowView/ProviderRowView.swift
@@ -8,54 +8,6 @@
 
 import SwiftUI
 
-struct ProviderRowViewModel {
-    let provider: Provider
-    let isDisabled: Bool
-    let badge: Badge?
-    let subtitles: [Subtitle]
-    let detailsType: DetailsType?
-    let tapAction: () -> Void
-}
-
-extension ProviderRowViewModel: Hashable, Identifiable {
-    var id: Int { hashValue }
-    static func == (lhs: ProviderRowViewModel, rhs: ProviderRowViewModel) -> Bool {
-        lhs.id == rhs.id
-    }
-
-    func hash(into hasher: inout Hasher) {
-        hasher.combine(provider)
-        hasher.combine(isDisabled)
-        hasher.combine(badge)
-        hasher.combine(subtitles)
-        hasher.combine(detailsType)
-    }
-}
-
-extension ProviderRowViewModel {
-    struct Provider: Hashable {
-        let iconURL: URL
-        let name: String
-        let type: String
-    }
-
-    enum Badge: String, Hashable {
-        case permissionNeeded
-        case bestRate
-    }
-
-    enum Subtitle: Hashable, Identifiable {
-        var id: Int { hashValue }
-
-        case text(String)
-        case percent(String, signType: ChangeSignType)
-    }
-
-    enum DetailsType: Hashable {
-        case selected
-        case chevron
-    }
-}
 
 struct ProviderRowView: View {
     let viewModel: ProviderRowViewModel
@@ -69,10 +21,7 @@ struct ProviderRowView: View {
     private var content: some View {
         HStack(spacing: 12) {
             IconView(url: viewModel.provider.iconURL, size: CGSize(bothDimensions: 36))
-                .cornerRadius(0)
-                .modifier(if: viewModel.isDisabled) {
-                    $0.saturation(0)
-                }
+                .saturation(viewModel.isDisabled ? 0 : 1)
 
             VStack(alignment: .leading, spacing: 4) {
                 titleView
