+++ Tangem/Resources/Localizations/it.lproj/Localizable.strings
@@ -181,7 +181,7 @@
 "exchange_receive_view_header" = "You Receive";
 "exchange_send_view_header" = "You Send";
 "express_choose_providers_subtitle" = "Providers facilitate transactions, ensuring smooth and efficient token exchanges";
-"express_choose_providers_title" = "Choose provider";
+"express_choose_providers_title" = "Choose Provider";
 "express_provider_best_rate" = "Best Rate";
 "express_provider_permission_needed" = "Permission Needed";
 "feedback_preface_rate_negative" = "Tell us what functions you are missing, and we will try to help you.";
